This seems to work; however, due to the nature of design within USB/LSB two blocks with the same seed can desync and resync randomly over the course of an observation (different sample request counts driven through gnuradio). While not a critical issue, it will cause.. instability under certain conditions. This is rooted within the use of variable sized FFT windows for the Hilbert transform, might need to be adjusted for repeatability.

+++ include/signal_exciter/random_signal_config.h
@@ -109,7 +109,7 @@ namespace gr {
       std::vector<float> weight;            // weights of each spectral guassian component (analog mods)
       size_t components;                    // number of spectral guassian components (analog mods)
 
-      float frac_symb_offset;               // Fractional Symbol Offset (pairs with interp)
+      double frac_symb_offset;               // Fractional Symbol Offset (pairs with interp)
     };
 
   } // namespace signal_exciter
